+++ CHANGELOG.md
@@ -27,7 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed
 
 -   `<MultiSuggestField />`
-    -   Added logic for contolled state for the component if `selectedItems` and `onSelection` passed.
+    -   Added logic for contolled state for the component if `selectedItems`, `prePopulateWithItems` and `onSelection` passed.
     -   Fixed deferred `selectedItems` setting.
 -   `<StickyNoteModal/>`
     -   static test id `data-test-id="sticky-note-modal"` will be removed with next major version